BasketItem için serviste findBasketItem save ve delete metotşarı tamamlandı

+++ src/main/java/org/example/eticaret/entity/BasketItem.java
@@ -12,7 +12,7 @@ import lombok.*;
 public class BasketItem {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long basketItemId;
+    private int basketItemId;
     private int count;
     private double basketItemTotalPrice;
